Build 20230531: Add docker.

+++ FILES/usr/bin/ppg.sh
@@ -36,26 +36,76 @@ fast_node_sel() {
         clash_web_password="clashpass"
     fi
     if [ -z "$test_node_url" ]; then
-        test_node_url="http://www.google.com"
+        test_node_url="https://www.youtube.com/generate_204"
     fi
     if [ -z "$ext_node" ]; then
         ext_node="Traffic|Expire| GB|Days|Date"
     fi
+    log "Try to switch the fastest node..." warn
     ppgw -apiurl="http://127.0.0.1:""$clash_web_port" -secret="$clash_web_password" -test_node_url="$test_node_url" -ext_node="$ext_node" >/dev/tty0
 }
-start_clash() {
+kill_cron() {
+    if ps | grep -v "grep" | grep "/etc/cron"; then
+        /etc/init.d/cron stop
+    fi
+}
+kill_clash_cache() {
+    if [ -f /etc/config/clash/cache.db ]; then
+        rm -f /etc/config/clash/cache.db
+    fi
+}
+kill_clash() {
+    if ps | grep -v "grep" | grep "/etc/config/clash"; then
+        kill $(pgrep -x "/usr/bin/clash")
+        nft flush ruleset
+    fi
+}
+load_clash() {
     if [ -f /tmp/clash.yaml ]; then
         if [ -f /tmp/ppgw.ini ]; then
             . /tmp/ppgw.ini 2>/dev/tty0
         fi
         if [ -z "$test_node_url" ]; then
-            test_node_url="http://www.google.com"
+            test_node_url="https://www.youtube.com/generate_204"
         fi
-        sed "s|http://www.google.com|$test_node_url|g" /etc/config/clash/clash-dashboard/index_base.html >/etc/config/clash/clash-dashboard/index.html
-        /usr/bin/clash -d /etc/config/clash -f /tmp/clash.yaml >/dev/tty0 2>&1 &
+        if [ -z "$clash_web_port" ]; then
+            clash_web_port="80"
+        fi
+        if [ -z "$clash_web_password" ]; then
+            clash_web_password="clashpass"
+        fi
+        sed "s|https://www.youtube.com/generate_204|$test_node_url|g" /etc/config/clash/clash-dashboard/index_base.html >/etc/config/clash/clash-dashboard/index.html
+        if ps | grep -v "grep" | grep "/etc/config/clash"; then
+            ppgw -reload -apiurl="http://127.0.0.1:""$clash_web_port" -secret="$clash_web_password" >/dev/tty0 2>&1
+        else
+            /usr/bin/clash -d /etc/config/clash -f /tmp/clash.yaml >/dev/tty0 2>&1 &
+        fi
     else
         log "The clash.yaml generation failed." warn
     fi
+
+    if ps | grep -v "grep" | grep "/etc/config/clash"; then
+        if [ "$1" = "yes" ]; then
+            fast_node_sel || fast_node_sel || fast_node_sel || fast_node_sel || fast_node_sel
+        fi
+        if [ "$2" = "no" ]; then
+            if nft list ruleset | grep "clashtcp"; then
+                log "[OK] nft rule TCP OK." succ
+
+            else
+                log "[ADD] Add nft rule TCP..." warn
+                /usr/bin/nft_tcp.sh
+            fi
+        else
+            if nft list ruleset | grep "clashboth"; then
+                log "[OK] nft rule TCP/UDP OK." succ
+            else
+                log "[ADD] Add nft rule TCP/UDP..." warn
+                /usr/bin/nft.sh
+            fi
+        fi
+    fi
+
 }
 gen_hash() {
     if [ -f /tmp/ppgw.ini ]; then
@@ -83,6 +133,14 @@ get_conf() {
     down_type=$2
     down_sdns=$3
     if [ "$down_type" = "ini" ]; then
+        if [ -f /www/ppgw.ini ]; then
+            if [ -f /tmp/ppgw.ini ]; then
+                log "Load local ppgw.ini" succ
+            else
+                cp /www/ppgw.ini /tmp/ppgw.ini
+            fi
+            return 0
+        fi
         file_down="/tmp/ppgw.ini.down"
     fi
     if [ "$down_type" = "yaml" ]; then
@@ -95,14 +153,24 @@ get_conf() {
     if [ -f /tmp/ppgw.ini ]; then
         . /tmp/ppgw.ini 2>/dev/tty0
     fi
+    if [ -z "$dns_ip" ]; then
+        dns_ip="1.0.0.1"
+    fi
+    if [ -z "$dns_port" ]; then
+        dns_port="53"
+    fi
     echo "127.0.0.1 localhost" >/etc/hosts
     if [ "$down_sdns" = "yes" ]; then
         genHost=$(ppgw -server "$dns_ip" -port "$dns_port" -rawURL "$down_url")
     else
         genHost=$(ppgw -rawURL "$down_url")
     fi
+    if [ "$?" = "1" ]; then
+        log "Nslookup failed." warn
+        return 1
+    fi
     echo "$genHost" >>/etc/hosts
-    wget --header="User-Agent: ClashforWindows/0.20.23" --timeout=10 --tries=1 --no-check-certificate "$down_url" -O "$file_down_tmp" >/dev/tty0 2>&1
+    ppgw -downURL "$down_url" -output "$file_down_tmp" >/dev/tty0 2>&1
     echo "127.0.0.1 localhost" >/etc/hosts
     if [ "$down_type" = "ini" ]; then
         if head -1 "$file_down_tmp" | grep -q "#paopao-gateway"; then
@@ -126,7 +194,18 @@ get_conf() {
     if [ "$down_type" = "yaml" ]; then
         if grep -q "proxies:" "$file_down_tmp"; then
             cp "$file_down_tmp" "$file_down"
-            sed 's/\r$//' "$file_down" | grep -v "\- RULE-SET" >"/tmp/paopao_custom.yaml"
+            if [ -f /www/ppgw.ini ]; then
+                . /www/ppgw.ini
+            fi
+            if [ "$fast_node" = "yes" ]; then
+                sed 's/\r$//' "$file_down" | grep -v "\- RULE-SET" | sed "s/rule-providers:/rule-disable-providers:/g" | sed "s/proxy-groups:/proxy-disable-groups:/g" | sed "s/rules:/ru-disable-les:/g" >"/tmp/paopao_custom.yaml"
+            else
+                if [ -f /www/clash_core ]; then
+                    sed 's/\r$//' "$file_down" >"/tmp/paopao_custom.yaml"
+                else
+                    sed 's/\r$//' "$file_down" | grep -v "\- RULE-SET" >"/tmp/paopao_custom.yaml"
+                fi
+            fi
             log "[Succ] Get ""$down_url" succ
             return 0
         fi
@@ -146,6 +225,24 @@ try_conf() {
     down_type=$2
     paopao="paopao.dns"
     log "Try to get new ""$conf"
+    if [ -f /www/ppgwurl.ini ] && [ "$down_type" = "ini" ]; then
+        . /www/ppgwurl.ini
+        if [ -n "$ppgwurl" ]; then
+            get_conf "$ppgwurl" "ini"
+            return 0
+        fi
+    fi
+
+    if [ -f /www/custom.yaml ] && [ "$down_type" = "yaml" ]; then
+        if [ -f "/tmp/ppgw.yaml.down" ] && [ -f "/tmp/paopao_custom.yaml" ]; then
+            log "Load local yaml" succ
+        else
+            cp /www/custom.yaml /tmp/ppgw.yaml.down
+            cp /www/custom.yaml /tmp/paopao_custom.yaml
+        fi
+        return 0
+    fi
+
     if [ -n "$try_succ_host" ]; then
         get_conf "http://""$try_succ_host":"$try_succ_confport""/""$conf" "$down_type"
     else
@@ -201,11 +298,7 @@ reload_gw() {
     if [ -z "$udp_enable" ]; then
         udp_enable="yes"
     fi
-    if [ "$udp_enable" = "no" ]; then
-        /usr/bin/nft_tcp.sh
-    else
-        /usr/bin/nft.sh
-    fi
+
     # route table
     if ip route list table 100 2>&1 | grep -q local; then
         log "[OK] table100 OK." succ
@@ -229,6 +322,9 @@ reload_gw() {
         dns_ip="1.0.0.1"
     fi
 
+    if [ -z "$dns_port" ]; then
+        dns_port="53"
+    fi
     if [ -z "$openport" ]; then
         openport="false"
     else
@@ -239,9 +335,6 @@ reload_gw() {
         fi
     fi
 
-    if [ -z "$dns_port" ]; then
-        dns_port="53"
-    fi
     if [ -z "$clash_web_port" ]; then
         clash_web_port="80"
     fi
@@ -251,7 +344,7 @@ reload_gw() {
     fi
 
     if [ -z "$mode" ]; then
-        mode="socks5"
+        mode="free"
     fi
 
     if [ -z "$socks5_ip" ]; then
@@ -278,7 +371,7 @@ reload_gw() {
         rm "/tmp/clash.yaml"
     fi
     if [ "$mode" = "socks5" ]; then
-        /etc/init.d/cron stop
+        kill_cron
         sed -i "s/{clashmode}/rule/g" /tmp/clash_base.yaml
         sed 's/\r$//' /etc/config/clash/socks5.yaml >/tmp/clash_socks5.yaml
         sed -i "s/{socks5_ip}/$socks5_ip/g" /tmp/clash_socks5.yaml
@@ -286,7 +379,7 @@ reload_gw() {
         ppgw -input /tmp/clash_socks5.yaml -input /tmp/clash_base.yaml -output /tmp/clash.yaml
     fi
     if [ "$mode" = "yaml" ]; then
-        /etc/init.d/cron stop
+        kill_cron
         try_conf "$yamlfile" "yaml"
     fi
     if [ "$mode" = "suburl" ]; then
@@ -305,29 +398,29 @@ reload_gw() {
             log "Bad suburl" warn
         fi
     fi
+
     if [ "$mode" = "yaml" ] || [ "$mode" = "suburl" ]; then
-        if grep -q "rules:" /tmp/paopao_custom.yaml; then
-            sed -i "s/{clashmode}/rule/g" /tmp/clash_base.yaml
-            export clashmode=rule
-        else
-            sed -i "s/{clashmode}/global/g" /tmp/clash_base.yaml
+        if [ "$fast_node" = "yes" ]; then
             export clashmode=global
+        else
+            if grep -q "rules:" /tmp/paopao_custom.yaml; then
+                export clashmode=rule
+            else
+                export clashmode=global
+            fi
         fi
         if [ -f /tmp/paopao_custom.yaml ]; then
+            sed -i "s/{clashmode}/$clashmode/g" /tmp/clash_base.yaml
             ppgw -input /tmp/paopao_custom.yaml -input /tmp/clash_base.yaml -output /tmp/clash.yaml
         fi
     fi
+
     if [ "$mode" = "free" ]; then
-        /etc/init.d/cron stop
+        kill_cron
         sed -i "s/{clashmode}/direct/g" /tmp/clash_base.yaml
         cat /tmp/clash_base.yaml >/tmp/clash.yaml
     fi
-    if ps | grep -v "grep" | grep clash; then
-        kill $(pgrep -x "/usr/bin/clash")
-    fi
-    log "Start clash..." warn
-    start_clash
-    sleep 3
+    log "Load clash config..." warn
     if [ -z "$fast_node" ]; then
         if [ "$clashmode" = "global" ]; then
             fast_node=yes
@@ -335,9 +428,8 @@ reload_gw() {
             fast_node=no
         fi
     fi
-    if [ "$fast_node" = "yes" ]; then
-        fast_node_sel
-    fi
+    load_clash $fast_node $udp_enable
+
 }
 
 if [ "$1" = "reload" ]; then
@@ -350,6 +442,7 @@ if [ "$1" = "cron" ]; then
         . /tmp/ppgw.ini 2>/dev/tty0
     fi
     get_conf "$suburl" "yaml" "yes"
+    reload_gw
     exit
 fi
 sysctl -w net.ipv6.conf.all.disable_ipv6=1
@@ -360,6 +453,15 @@ echo " " >/dev/tty0
 last_hash="empty"
 last_yaml_hash="empty"
 while true; do
+    if [ -f /tmp/ppgw.ini ]; then
+        . /tmp/ppgw.ini 2>/dev/tty0
+        old_fake_cidr=$fake_cidr
+        old_openport=$openport
+        old_clash_web_port=$clash_web_port
+        old_clash_web_password=$clash_web_password
+        old_suburl=$suburl
+        old_fast_node=$fast_node
+    fi
     try_conf "ppgw.ini" "ini"
     hash=$(gen_hash)
     log "[OLD PPGW HASH]: ""$last_hash"
@@ -367,8 +469,37 @@ while true; do
     if [ "$hash" != "$last_hash" ]; then
         if echo "$hash" | grep -Eqo "[a-z0-9]{32}"; then
             log "The hash has changed, reload gateway." warn
+            if [ -f /tmp/ppgw.ini ]; then
+                . /tmp/ppgw.ini 2>/dev/tty0
+            fi
+            if [ "$old_fake_cidr" != "$fake_cidr" ]; then
+                kill_clash
+                kill_clash_cache
+            fi
+            if [ "$old_openport" != "$openport" ]; then
+                kill_clash
+            fi
+            if [ "$old_clash_web_port" != "$clash_web_port" ]; then
+                kill_clash
+            fi
+            if [ "$old_clash_web_password" != "$clash_web_password" ]; then
+                kill_clash
+            fi
+            if [ "$old_fast_node" != "$fast_node" ]; then
+                kill_clash
+                if [ "$mode" = "suburl" ]; then
+                    get_conf "$suburl" "yaml" "yes"
+                fi
+            fi
+            if [ "$mode" = "suburl" ]; then
+                if [ "$old_suburl" != "$suburl" ]; then
+                    get_conf "$suburl" "yaml" "yes"
+                fi
+            fi
             reload_gw
-            last_hash="$hash"
+            if [ -f /tmp/ppgw.ini ]; then
+                last_hash=$(gen_hash)
+            fi
             continue
         fi
     fi
@@ -385,17 +516,11 @@ while true; do
         if [ "$last_yaml_hash" != "$yaml_hash" ]; then
             log "The yaml hash has changed, reload gateway." warn
             reload_gw
-            last_yaml_hash="$yaml_hash"
+            if [ -f /tmp/ppgw.yaml.down ]; then
+                last_yaml_hash=$(gen_yaml_hash "/tmp/ppgw.yaml.down")
+            fi
             continue
         fi
-    fi
-
-    if ps | grep -v "grep" | grep "clash"; then
-        echo "Clash running OK."
-    else
-        ps | grep clash >/dev/tty0 &
-        log "Try to run Clash again..." warn
-        start_clash
     fi
     if [ -z "$fast_node" ]; then
         if [ "$clashmode" = "global" ]; then
@@ -404,22 +529,30 @@ while true; do
             fast_node=no
         fi
     fi
+    if ps | grep -v "grep" | grep "/etc/config/clash"; then
+        echo "Clash running OK."
+    else
+        ps | grep clash >/dev/tty0 &
+        log "Try to run Clash again..." warn
+        load_clash $fast_node $udp_enable
+    fi
 
     if [ "$fast_node" = "yes" ] || [ "$fast_node" = "check" ]; then
         if [ -z "$test_node_url" ]; then
-            test_node_url="http://www.google.com"
+            test_node_url="https://www.youtube.com/generate_204"
         fi
         proxytest=$(ppgw -testProxy http://127.0.0.1:1080 -test_node_url "$test_node_url")
         if [ $? -eq 0 ]; then
             log "$proxytest" succ
         else
             log "Node Check Fail:""$proxytest" warn
+            kill_clash
             if [ "$fast_node" = "yes" ]; then
-                log "Try to switch the fastest node..." warn
-                fast_node_sel
-            else
                 log "Try to update and reload..." warn
-                /usr/bin/reload
+                if [ "$mode" = "suburl" ]; then
+                    get_conf "$suburl" "yaml" "yes"
+                fi
+                reload_gw
             fi
         fi
     fi
